+++ docs/changelog.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 0.3.14
 
 ### Added
--
+- In API, suppress specific scanners when running the analysis.
 
 ### Fixed
 -